+++ src/main/java/com/tsp/new_tsp_front/api/agency/service/FrontAgencyJpaService.java
@@ -1,15 +1,16 @@
 package com.tsp.new_tsp_front.api.agency.service;
 
 import com.tsp.new_tsp_front.api.agency.domain.FrontAgencyDTO;
+import com.tsp.new_tsp_front.api.agency.domain.FrontAgencyEntity;
+import com.tsp.new_tsp_front.api.agency.service.impl.FrontAgencyJpaQueryRepository;
 import com.tsp.new_tsp_front.api.agency.service.impl.FrontAgencyJpaRepository;
 import com.tsp.new_tsp_front.exception.TspException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.cache.annotation.CachePut;
-import org.springframework.cache.annotation.Cacheable;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Map;
 
 import static com.tsp.new_tsp_front.exception.ApiExceptionType.*;
@@ -17,20 +18,12 @@ import static com.tsp.new_tsp_front.exception.ApiExceptionType.*;
 @Service
 @RequiredArgsConstructor
 public class FrontAgencyJpaService {
+    private final FrontAgencyJpaQueryRepository frontAgencyJpaQueryRepository;
     private final FrontAgencyJpaRepository frontAgencyJpaRepository;
 
-    /**
-     * <pre>
-     * 1. MethodName : findAgencyCount
-     * 2. ClassName  : FrontAgencyJpaService.java
-     * 3. Comment    : 프론트 > Agency 리스트 갯수 조회
-     * 4. 작성자      : CHO
-     * 5. 작성일      : 2022. 08. 24.
-     * </pre>
-     */
-    @Transactional(readOnly = true)
-    public int findAgencyCount(Map<String, Object> agencyMap) {
-        return frontAgencyJpaRepository.findAgencyCount(agencyMap);
+    private FrontAgencyEntity oneAgency(Long idx) {
+        return frontAgencyJpaRepository.findById(idx)
+                .orElseThrow(() -> new TspException(NOT_FOUND_AGENCY));
     }
 
     /**
@@ -42,10 +35,9 @@ public class FrontAgencyJpaService {
      * 5. 작성일      : 2022. 08. 24.
      * </pre>
      */
-    @Cacheable(value = "agency", key = "#agencyMap")
     @Transactional(readOnly = true)
-    public List<FrontAgencyDTO> findAgencyList(Map<String, Object> agencyMap) {
-        return frontAgencyJpaRepository.findAgencyList(agencyMap);
+    public Page<FrontAgencyDTO> findAgencyList(Map<String, Object> agencyMap, PageRequest pageRequest) {
+        return frontAgencyJpaQueryRepository.findAgencyList(agencyMap, pageRequest);
     }
 
     /**
@@ -57,10 +49,9 @@ public class FrontAgencyJpaService {
      * 5. 작성일      : 2022. 08. 24.
      * </pre>
      */
-    @Cacheable(value = "agency", key = "#idx")
     @Transactional(readOnly = true)
     public FrontAgencyDTO findOneAgency(Long idx) {
-        return this.frontAgencyJpaRepository.findOneAgency(idx);
+        return FrontAgencyEntity.toDto(oneAgency(idx));
     }
 
     /**
@@ -72,10 +63,9 @@ public class FrontAgencyJpaService {
      * 5. 작성일      : 2022. 09. 17.
      * </pre>
      */
-    @Cacheable(value = "agency", key = "#idx")
     @Transactional(readOnly = true)
     public FrontAgencyDTO findPrevOneAgency(Long idx) {
-        return this.frontAgencyJpaRepository.findPrevOneAgency(idx);
+        return this.frontAgencyJpaQueryRepository.findPrevOneAgency(idx);
     }
 
     /**
@@ -87,11 +77,10 @@ public class FrontAgencyJpaService {
      * 5. 작성일      : 2022. 09. 17.
      * </pre>
      */
-    @Cacheable(value = "agency", key = "#idx")
     @Transactional(readOnly = true)
     public FrontAgencyDTO findNextOneAgency(Long idx) {
         try {
-            return this.frontAgencyJpaRepository.findNextOneAgency(idx);
+            return this.frontAgencyJpaQueryRepository.findNextOneAgency(idx);
         } catch (Exception e) {
             throw new TspException(NOT_FOUND_AGENCY);
         }
@@ -106,11 +95,10 @@ public class FrontAgencyJpaService {
      * 5. 작성일      : 2022. 08. 24.
      * </pre>
      */
-    @CachePut(value = "agency", key = "#idx")
     @Transactional
     public int favoriteAgency(Long idx) {
         try {
-            return frontAgencyJpaRepository.favoriteAgency(idx);
+            return frontAgencyJpaQueryRepository.favoriteAgency(idx);
         } catch (Exception e) {
             throw new TspException(ERROR_AGENCY_LIKE);
         }
